% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{areapl}
\alias{areapl}
\title{Calculate area of polynomial}
\usage{
areapl(poly)
}
\arguments{
\item{poly}{- matrix describing polynomial}
}
\value{
W - area of polynomial
}
\description{
Calculate area of polynomial
}
