% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intensity_temporal}
\alias{intensity_temporal}
\title{Calculate intensity function for temporal hawkes}
\usage{
intensity_temporal(mu, alpha, beta, times, evalpt)
}
\arguments{
\item{mu}{- background parameter}

\item{alpha}{- alpha parameter}

\item{beta}{- beta parameter}

\item{times}{- history of previous times}

\item{evalpt}{- point to evaluate}
}
\value{
lambda - intenisty at evalpt
}
\description{
Calculate intensity function for temporal hawkes
}
