% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{pip}
\alias{pip}
\title{Point in polygon}
\usage{
pip(x, y, poly)
}
\arguments{
\item{x}{- vector of x positions}

\item{y}{- vector of y positions}

\item{poly}{- matrix defining polygon (\eqn{N} x \eqn{2})}
}
\value{
A list containing the x and y coordinates of the points inside the polygon
}
\description{
Determines if a point is in a polygon or on a polygon boundary
}
