% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ptinpoly}
\alias{ptinpoly}
\title{Calculate if points are in the polynomial}
\usage{
ptinpoly(x, y, xp, yp, bb)
}
\arguments{
\item{x}{- vector of x coordinates}

\item{y}{- vector of y coordinates}

\item{xp}{- vector of x coordinates of polynomial}

\item{yp}{- vector of y coordinates of polynomial}

\item{bb}{- matrix of bounding box of polynomial}
}
\value{
inout - vector of 1 if point is in polynomial and 0 if not
}
\description{
Calculate if points are in the polynomial
}
