% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{stpp.mle}
\alias{stpp.mle}
\title{MLE Estimation of Spatio-Temporal Hawkes Model Parameters}
\usage{
stpp.mle(data, poly, t_max = max(data$t), initval = NA, print = TRUE)
}
\arguments{
\item{data}{- A DataFrame containing \eqn{x},\eqn{y}, and \eqn{t}}

\item{poly}{- a matrix defining the polygon}

\item{t_max}{- maximum time value (default = max(times))}

\item{initval}{- vector of two elements describing missing time range (default = NA)}

\item{print}{- print progress (default = TRUE)}
}
\value{
A list containing the parameter values and likelihood value
}
\description{
Maximum likelihood estimation of the parameters of a spatio-temporal exponential decay Hawkes model.
}
