% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{temporal.catmark.mle}
\alias{temporal.catmark.mle}
\title{MLE Estimation of Temporal Hawkes Model Parameters with Categorical Marks}
\usage{
temporal.catmark.mle(t, marks, t_max = max(t), initval = NA, print = TRUE)
}
\arguments{
\item{t}{- vector of arrival times}

\item{marks}{- vector of marks}

\item{t_max}{- maximum time value (default = max(times))}

\item{initval}{- initial parameter values for likelihood optimization}

\item{print}{- print progress (default = TRUE)}
}
\value{
A list containing the parameter values and likelihood value
}
\description{
Maximum likelihood estimation of the parameters of a temporal exponential decay Hawkes model
}
