% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{temporal.mle}
\alias{temporal.mle}
\title{MLE Estimation of Temporal Hawkes Model Parameters}
\usage{
temporal.mle(t, t_max = max(t), initval = NA, print = TRUE)
}
\arguments{
\item{t}{- vector of arrival times}

\item{t_max}{- maximum time value (default = max(times))}

\item{initval}{- vector of two elements describing missing time range (default = NA)}

\item{print}{- print progress (default = TRUE)}
}
\value{
A list containing the parameter values and likelihood value
}
\description{
Maximum likelihood estimation of the parameters of a temporal exponential decay Hawkes model
}
