% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpphawkes-package.R
\docType{package}
\name{stpphawkes}
\alias{stpphawkes}
\alias{stpphawkes-package}
\title{Marked Hawkes Process with Missing Data}
\description{
A library for estimation of spatio-temporal Hawkes process parameters with
missing data support
}
\references{
J. D. Tucker, L. Shand, and J. R. Lewis, “Handling Missing Data in Self-Exciting Point Process Models,”
  Spatial Statistics, vol. 29. pp. 160-176, 2019.
}
